Fix: 사운드오류 ETC
