--- Assets/03.Scripts/IntroCamAction.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: ac527a36261697344a9639553b891918
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
