--- Assets/03.Scripts/MBTIInfo.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 9c79bf3cd105e9c4fae2bef8a78aad8c
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
