+++ Assets/03.Scripts/SceneBaseScripts/ChatScene.cs
@@ -2,7 +2,7 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
-public class IntroScene : MonoBehaviour
+public class ChatScene : MonoBehaviour
 {
     // Start is called before the first frame update
     void Start()
