+++ Assets/03.Scripts/SceneBaseScripts/MainScene.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 781d5892c815cef4da53207c48123c0b
+guid: 9c79bf3cd105e9c4fae2bef8a78aad8c
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
