+++ Assets/03.Scripts/SceneBaseScripts/MenuScene.cs
@@ -7,13 +7,12 @@ using UnityEngine.UI;
 public class MenuScene : MonoBehaviour
 {
     public Button checkBtn;
-    public GameObject fireWriteManager;
     private string userMbti;
     
     //03.MenuScene -> MBTI 검사시작 또는 mbti 아는 경우 로직 처리 스크립트
     private List<string> validMbtiList = new List<string> {
-        "istj", "isfp", "infj", "intj", "infp", "intp", "isfj", "istp",
-        "estp", "esfp", "enfp", "entp", "estj", "esfj", "enfj", "entj"
+        "ISTJ", "ISFP", "INFJ", "INTJ", "INFP", "INTP", "ISFJ", "ISTP",
+        "ESTP", "ESFP", "ENFP", "ENTP", "ESTJ", "ESFJ", "ENFJ", "ENTJ"
     };
     //사용자 MBTI 입력 감지 호출 함수
     public void OnInputFieldEndEdit(string str)
@@ -23,8 +22,8 @@ public class MenuScene : MonoBehaviour
         {
             checkBtn.GetComponent<Button>().interactable = true;
             //유효한 MBTI
-            userMbti = str;
-            fireWriteManager.GetComponent<FirebaseWriteManager>().SaveMBTI(userMbti);
+            userMbti = str.ToUpper();
+            FirebaseWriteManager.Instance.SaveMBTI(userMbti);
         }
         else
         {
@@ -37,7 +36,7 @@ public class MenuScene : MonoBehaviour
     private bool IsValidMbti(string mbti)
     {
         //입력 MBTI가 유효한 MBTI목록에 포함되는지 확인
-        return validMbtiList.Contains(mbti.ToLower());
+        return validMbtiList.Contains(mbti.ToUpper());
     }
     
 }
