--- Assets/03.Scripts/TestResult.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 4ed75f074eddf6c42ae5bb911f6be864
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
