+++ Assets/05.Scenes/PC/04_02.ContentScene.unity
@@ -3315,7 +3315,7 @@ GameObject:
   - component: {fileID: 1388732047}
   - component: {fileID: 1388732046}
   m_Layer: 0
-  m_Name: UIManager
+  m_Name: SceneCtrlManager
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
